Former-commit-id: 802fd48e641bc7741bdc6286d0d6b13129a5eb16
Former-commit-id: 663910d31dc2197aaae953f948b59f4bfb17bd55

+++ src/pages/phone_pages/PhoneChatEdit.jsx
@@ -23,7 +23,7 @@ function PhoneChatEdit() {
   function changeName() {
     API.put(`/chat/room`, JSON.stringify(content))
       .then((res) => {
-        console.log(res);
+        // console.log(res);
       })
       .then(() => navigation(`../chat/room/${room}`));
   }
@@ -35,23 +35,25 @@ function PhoneChatEdit() {
   }
 
   const chatOut = async () => {
-    const data = {
-      croom_no: room,
-      user_id: user,
-    };
+    if (window.confirm('채팅방을 나가시겠습니까?') === true) {
+      const data = {
+        croom_no: room,
+        user_id: user,
+      };
 
-    // not_read 최신화
-    await API.put('/chat', JSON.stringify(data)).then((res) => {});
+      // not_read 최신화
+      await API.put('/chat', JSON.stringify(data)).then((res) => {});
 
-    // chat 나가기
-    await API.delete(`/participant`, { data: data })
-      .then((res) => {
-        console.log(res);
-      })
-      .catch((err) => {
-        console.log(err);
-      });
-    navigation(`../chat`);
+      // chat 나가기
+      await API.delete(`/participant`, { data: data })
+        .then((res) => {
+          // console.log(res);
+        })
+        .catch((err) => {
+          // console.log(err);
+        });
+      navigation(`../chat`);
+    }
   };
 
   return (
